requirements.txt
pdf2image
contextlib2
object_detection
protobuf
tensorflow>=2.8.0
numpy>=1.19.2
Pillow>=8.1.0
scipy
wsl
six
google-api-python-client>=1.6.7
kaggle>=1.3.9
numpy>=1.20
oauth2client
pandas>=0.22.0
psutil>=5.4.3
py-cpuinfo>=3.3.0
scipy>=0.19.1
tensorflow-hub>=0.6.0
tensorflow-model-optimization>=0.4.1
tensorflow-datasets
tf-keras>=2.16.0
gin-config
tf_slim>=1.1.0
Cython
matplotlib
# Loader becomes a required positional argument in 6.0 in yaml.load
pyyaml>=6.0.0
# CV related dependencies
opencv-python-headless
Pillow
pycocotools
# NLP related dependencies
seqeval
sentencepiece
sacrebleu
# Projects/vit dependencies
immutabledict
